requirements.txt
pandas~=2.2.3
psycopg2-binary~=2.9.10
SQLAlchemy~=2.0.36
pysftp~=0.2.9